requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aiohttp==3.8.4
    # via
    #   -r requirements.in
    #   langchain
    #   openai
aiosignal==1.3.1
    # via aiohttp
anyio==3.6.2
    # via
    #   starlette
    #   watchfiles
art==5.9
    # via -r requirements.in
async-timeout==4.0.2
    # via
    #   aiohttp
    #   langchain
attrs==23.1.0
    # via aiohttp
backoff==2.2.1
    # via posthog
certifi==2022.12.7
    # via
    #   clickhouse-connect
    #   requests
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
chromadb==0.3.21
    # via -r requirements.in
click==8.1.3
    # via
    #   nltk
    #   uvicorn
clickhouse-connect==0.5.21
    # via chromadb
contourpy==1.0.7
    # via matplotlib
cycler==0.11.0
    # via matplotlib
dataclasses-json==0.5.7
    # via langchain
dnspython==2.3.0
    # via pinecone-client
duckdb==0.7.1
    # via chromadb
easyocr==1.6.2
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
fastapi==0.95.1
    # via chromadb
filelock==3.12.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fonttools==4.39.3
    # via matplotlib
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
fsspec==2023.4.0
    # via huggingface-hub
h11==0.14.0
    # via uvicorn
hnswlib==0.7.0
    # via chromadb
httptools==0.5.0
    # via uvicorn
huggingface-hub==0.14.1
    # via
    #   sentence-transformers
    #   transformers
idna==3.4
    # via
    #   anyio
    #   requests
    #   yarl
imageio==2.28.0
    # via scikit-image
jinja2==3.1.2
    # via torch
joblib==1.2.0
    # via
    #   nltk
    #   scikit-learn
kiwisolver==1.4.4
    # via matplotlib
langchain==0.0.149
    # via -r requirements.in
lazy-loader==0.2
    # via scikit-image
loguru==0.7.0
    # via pinecone-client
lz4==4.3.2
    # via clickhouse-connect
markdown-it-py==2.2.0
    # via rich
markupsafe==2.1.2
    # via jinja2
marshmallow==3.19.0
    # via
    #   dataclasses-json
    #   marshmallow-enum
marshmallow-enum==1.5.1
    # via dataclasses-json
matplotlib==3.7.1
    # via openai
mdurl==0.1.2
    # via markdown-it-py
monotonic==1.6
    # via posthog
mpmath==1.3.0
    # via sympy
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
networkx==3.1
    # via
    #   scikit-image
    #   torch
ninja==1.11.1
    # via easyocr
nltk==3.8.1
    # via sentence-transformers
numexpr==2.8.4
    # via langchain
numpy==1.24.3
    # via
    #   chromadb
    #   contourpy
    #   easyocr
    #   hnswlib
    #   imageio
    #   langchain
    #   matplotlib
    #   numexpr
    #   openai
    #   opencv-python-headless
    #   pandas
    #   pinecone-client
    #   pywavelets
    #   scikit-image
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   shapely
    #   tifffile
    #   torchvision
    #   transformers
openai[embeddings]==0.27.4
    # via -r requirements.in
openapi-schema-pydantic==1.2.4
    # via langchain
opencv-python-headless==4.5.4.60
    # via easyocr
openpyxl==3.1.2
    # via openai
packaging==23.1
    # via
    #   huggingface-hub
    #   marshmallow
    #   matplotlib
    #   plotly
    #   scikit-image
    #   transformers
pandas==2.0.1
    # via
    #   chromadb
    #   openai
pandas-stubs==2.0.0.230412
    # via openai
pdf2image==1.16.3
    # via -r requirements.in
pillow==9.5.0
    # via
    #   easyocr
    #   imageio
    #   matplotlib
    #   pdf2image
    #   scikit-image
    #   torchvision
pinecone-client==2.2.1
    # via -r requirements.in
plotly==5.14.1
    # via openai
posthog==3.0.1
    # via chromadb
pyclipper==1.3.0.post4
    # via easyocr
pydantic==1.10.7
    # via
    #   chromadb
    #   fastapi
    #   langchain
    #   openapi-schema-pydantic
pygments==2.15.1
    # via rich
pyparsing==3.0.9
    # via matplotlib
python-bidi==0.4.2
    # via easyocr
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
    #   pinecone-client
    #   posthog
python-dotenv==1.0.0
    # via
    #   -r requirements.in
    #   uvicorn
pytz==2023.3
    # via
    #   clickhouse-connect
    #   pandas
pywavelets==1.4.1
    # via scikit-image
pyyaml==6.0
    # via
    #   easyocr
    #   huggingface-hub
    #   langchain
    #   pinecone-client
    #   transformers
    #   uvicorn
regex==2023.3.23
    # via
    #   nltk
    #   tiktoken
    #   transformers
requests==2.28.2
    # via
    #   chromadb
    #   huggingface-hub
    #   langchain
    #   openai
    #   pinecone-client
    #   posthog
    #   tiktoken
    #   torchvision
    #   transformers
rich==13.3.5
    # via -r requirements.in
scikit-image==0.20.0
    # via easyocr
scikit-learn==1.2.2
    # via
    #   openai
    #   sentence-transformers
scipy==1.10.1
    # via
    #   easyocr
    #   openai
    #   scikit-image
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==2.2.2
    # via chromadb
sentencepiece==0.1.98
    # via sentence-transformers
shapely==2.0.1
    # via easyocr
six==1.16.0
    # via
    #   posthog
    #   python-bidi
    #   python-dateutil
sniffio==1.3.0
    # via anyio
sqlalchemy==2.0.10
    # via langchain
starlette==0.26.1
    # via fastapi
sympy==1.11.1
    # via torch
tenacity==8.2.2
    # via
    #   langchain
    #   openai
    #   plotly
threadpoolctl==3.1.0
    # via scikit-learn
tifffile==2023.4.12
    # via scikit-image
tiktoken==0.3.3
    # via -r requirements.in
tokenizers==0.13.3
    # via transformers
torch==2.0.0
    # via
    #   easyocr
    #   sentence-transformers
    #   torchvision
torchvision==0.15.1
    # via
    #   easyocr
    #   sentence-transformers
tqdm==4.65.0
    # via
    #   huggingface-hub
    #   langchain
    #   nltk
    #   openai
    #   pinecone-client
    #   sentence-transformers
    #   transformers
transformers==4.28.1
    # via sentence-transformers
types-pytz==2023.3.0.0
    # via pandas-stubs
typing-extensions==4.5.0
    # via
    #   huggingface-hub
    #   pinecone-client
    #   pydantic
    #   sqlalchemy
    #   torch
    #   typing-inspect
typing-inspect==0.8.0
    # via dataclasses-json
tzdata==2023.3
    # via pandas
urllib3==1.26.15
    # via
    #   clickhouse-connect
    #   pinecone-client
    #   requests
uvicorn[standard]==0.21.1
    # via chromadb
uvloop==0.17.0
    # via uvicorn
watchfiles==0.19.0
    # via uvicorn
websockets==11.0.2
    # via uvicorn
yarl==1.9.2
    # via aiohttp
zstandard==0.21.0
    # via clickhouse-connect
